simple fifostack
using a slice as fifo-stack by appending for push and removing the first element for pop is horribly ineffecient, because the slice will move over the underlying array. So we frequently need to allocate a new underlying array and copy the whole thing.

Instead of doing that I implemented a simple custom type that does what we need a lot faster.