Create generalOptions.go
Some option presets I settled on after experimenting with BB(4,2).

I tried to get a set of heuristics which can be used in sequence to find different types of machines, with some increasing parameters linked together by a level, so that the same options can be repeated at higher level to find more complicated machines at the cost of more computation.

The result works for BB(4,2), requiring one of the level 4 options.

If these options are any good for running on pre-filtered tm lists remains to be seen. For example the heuristic for cycler variants is almost certain to be useless there.